requirements/local.txt
-r base.txt # includes the base.txt requirements file
django-debug-toolbar==1.6
sqlparse==0.2.2